+++ tw-tasks-kafka-listener/src/main/java/com/transferwise/tasks/helpers/kafka/IKafkaListenerConsumerPropertiesProvider.java
@@ -1,0 +1,8 @@
+package com.transferwise.tasks.helpers.kafka;
+
+import java.util.Map;
+
+public interface IKafkaListenerConsumerPropertiesProvider {
+
+  Map<String, Object> getProperties(int shard);
+}