+++ tw-tasks-kafka-listener/src/main/java/com/transferwise/tasks/helpers/kafka/meters/KafkaListenerMetricsTemplate.java
@@ -4,6 +4,10 @@ import static com.transferwise.tasks.helpers.ICoreMetricsTemplate.METRIC_PREFIX;
 
 import com.transferwise.common.baseutils.meters.cache.IMeterCache;
 import com.transferwise.common.baseutils.meters.cache.TagsSet;
+import io.micrometer.core.instrument.ImmutableTag;
+import io.micrometer.core.instrument.binder.kafka.KafkaClientMetrics;
+import java.util.List;
+import org.apache.kafka.clients.consumer.Consumer;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class KafkaListenerMetricsTemplate implements IKafkaListenerMetricsTemplate {
@@ -13,6 +17,7 @@ public class KafkaListenerMetricsTemplate implements IKafkaListenerMetricsTempla
 
   private static final String TAG_TOPIC = "topic";
   private static final String TAG_SHARD = "shard";
+  private static long kafkaConsumerId = 0;
 
   @Autowired
   private IMeterCache meterCache;
@@ -27,5 +32,16 @@ public class KafkaListenerMetricsTemplate implements IKafkaListenerMetricsTempla
   public void registerFailedCommit() {
     meterCache.counter(METRIC_CONSISTENT_KAFKA_CONSUMER_FAILED_COMMITS_COUNT, TagsSet.empty())
         .increment();
+  }
+
+  @Override
+  @SuppressWarnings("rawtypes")
+  public synchronized void registerKafkaConsumer(Consumer consumer, long shard) {
+    // Spring application are setting the tag `spring.id`, so we need to set it as well.
+
+    // Add specific client id for consumer.
+
+    new KafkaClientMetrics(consumer, List.of(new ImmutableTag("spring.id", "tw-tasks-kafka-listener-" + shard + "-" + (kafkaConsumerId++))))
+        .bindTo(meterCache.getMeterRegistry());
   }
 }